(refactor)[controllers]: group all decorators bound closely to controllers
