+++ src/app/types/argument-indices.type.ts
@@ -1,3 +1,3 @@
-import { MethodArgumentMetadataKey } from '../../method-arguments/method-arguments.constants';
+import { MethodArgumentMetadataKey } from '../../controllers/method-arguments/method-arguments.constants';
 
 export type ArgumentIndices = { [P in MethodArgumentMetadataKey]?: number };
