+++ src/controllers/index.ts
@@ -1,0 +1,3 @@
+export * from './controller';
+export * from './http-methods';
+export * from './method-arguments';
