+++ src/index.ts
@@ -1,7 +1,5 @@
 export * from './app';
-export * from './controller';
+export * from './controllers';
 export * from './dependency-injection';
-export * from './http-methods';
 export * from './logger';
-export * from './method-arguments';
 export * from './middleware';
